chore: split into submodules
